+++ gems/aws-sdk-s3/spec/plugins/checksum_algorithm_spec.rb
@@ -111,7 +111,7 @@ module Aws
         end
       end
 
-      context 'trailer checksum' do
+      context 'request trailer checksum' do
         it 'sets aws-chunked when no existing Content-Encoding header' do
           resp = client.put_object(bucket: bucket, key: key, body: body)
           expect(resp.context.http_request.headers['Content-Encoding']).to eq('aws-chunked')