+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased Changes
 ------------------
 
-* Issue - Preserve existing Content-Encoding when applying trailer checksum.
+* Issue - Preserve existing Content-Encoding when applying request trailer checksum.
 
 3.241.1 (2026-01-06)
 ------------------