+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 Unreleased Changes
 ------------------
 
+3.241.2 (2026-01-07)
+------------------
+
 * Issue - Preserve existing Content-Encoding when applying request trailer checksum.
 
 3.241.1 (2026-01-06)